Inline `firezone-gui-client-common`
