+++ rust/Cargo.toml
@@ -14,7 +14,6 @@ members = [
   "dns-types",
   "etherparse-ext",
   "gateway",
-  "gui-client/src-common",
   "gui-client/src-tauri",
   "headless-client",
   "ip-packet",
@@ -73,7 +72,6 @@ env_logger = "0.11.6"
 etherparse = { version = "0.17", default-features = false }
 etherparse-ext = { path = "etherparse-ext" }
 firezone-bin-shared = { path = "bin-shared" }
-firezone-gui-client-common = { path = "gui-client/src-common" }
 firezone-headless-client = { path = "headless-client" }
 firezone-logging = { path = "logging" }
 firezone-relay = { path = "relay/server" }
