--- rust/gui-client/src-common/Cargo.toml
@@ -1,51 +1,0 @@
-[package]
-name = "firezone-gui-client-common"
-# mark:next-gui-version
-version = "1.4.13"
-edition = { workspace = true }
-license = { workspace = true }
-
-[dependencies]
-anyhow = { workspace = true }
-arboard = { workspace = true }
-atomicwrites = { workspace = true }
-connlib-model = { workspace = true }
-derive_more = { workspace = true, features = ["debug"] }
-firezone-bin-shared = { workspace = true }
-firezone-headless-client = { workspace = true }
-firezone-logging = { workspace = true }
-firezone-telemetry = { workspace = true }
-futures = { workspace = true }
-hex = { workspace = true }
-keyring = { workspace = true, features = ["crypto-rust", "sync-secret-service", "windows-native"] }
-native-dialog = { workspace = true }
-output_vt100 = { workspace = true }
-png = { workspace = true } # `png` is mostly free since we already need it for Tauri
-rand = { workspace = true }
-reqwest = { workspace = true, features = ["stream", "rustls-tls"] }
-sadness-generator = { workspace = true }
-secrecy = { workspace = true }
-semver = { workspace = true, features = ["serde"] }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-subtle = { workspace = true }
-thiserror = { workspace = true }
-time = { workspace = true, features = ["formatting"] }
-tokio = { workspace = true }
-tokio-stream = { workspace = true }
-tracing = { workspace = true }
-tracing-log = { workspace = true }
-tracing-subscriber = { workspace = true, features = ["env-filter"] }
-url = { workspace = true }
-uuid = { workspace = true, features = ["v4"] }
-zip = { workspace = true, features = ["deflate", "time"] }
-
-[target.'cfg(target_os = "linux")'.dependencies]
-dirs = { workspace = true }
-tracing-journald = { workspace = true }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-winreg = { workspace = true }
-
-[lints]
-workspace = true
