+++ rust/gui-client/src-tauri/src/client/deep_link/linux.rs
