+++ rust/gui-client/src-tauri/src/client/gui/system_tray/compositor.rs
