+++ rust/gui-client/src-tauri/src/client/ipc.rs
