+++ rust/gui-client/src-tauri/src/main.rs
@@ -5,6 +5,13 @@
 #![cfg_attr(test, allow(clippy::unwrap_used))]
 
 mod client;
+mod uptime;
+
+/// The Sentry "release" we are part of.
+///
+/// IPC service and GUI client are always bundled into a single release.
+/// Hence, we have a single constant for IPC service and GUI client.
+const RELEASE: &str = concat!("gui-client@", env!("CARGO_PKG_VERSION"));
 
 fn main() -> anyhow::Result<()> {
     // Mitigates a bug in Ubuntu 22.04 - Under Wayland, some features of the window decorations like minimizing, closing the windows, etc., doesn't work unless you double-click the titlebar first.
